[kotlin] Add opt-in markers to extracted functions when necessary
If the extracted code references an experimental API, and the function
container is annotated with corresponding propagating opt-in markers or
the `@OptIn` annotation, these annotations should be added to the
generated function declaration as well. Without these annotations,
the updated code may stop compiling or would start to cause compiler
warnings.

This commit adds opt-in annotation tracking to 'Extract Function'
action: the minimal necessary subset of original annotations is
added to the extracted function. Propagating markers are generated if
original markers were propagating, and `@OptIn` annotation is generated
for non-propagating opted-in markers.

^KTIJ-15823 Fixed

GitOrigin-RevId: 2deddda57a8c9412f2b32fc06a6bde98a60015bc