[kotlin] Fix AddModifierFix creation in case of old FE
The actual fix is now method AddModifierFix.Factory.createModifierFix actually used.
Previously all the time AddModifierFix was created from AddModifierFix.Factory methods.

AddModifierFixMpp renamed to AddModifierFixFE10 because previous name is really misleading.
AddModifierFix should be aware of MPP all the time, but this functionality is supported
only in the old FE right now, so it isn't correct to use the AddModifierFix in old plugin

GitOrigin-RevId: f29a93b690d55389f57f92689bca447dacc61350